chore(executor): extract atomic + tx modules (#116 step 4) (#514)
Pulls the transaction primitives out into their own files:

- tx.rs (136 LOC) — `PoolTx<'a>` enum + `transaction_pool`
  user-facing helper. Per-backend `BEGIN` wrapper that hands the
  caller a backend-tagged transaction handle.
- atomic.rs (178 LOC) — Django-shape `atomic(&pool, |tx| ...)` +
  `on_commit(callback)` + `on_commit_pending()` (issue #44). Uses
  a `tokio::task_local!` to queue commit-time callbacks.

The `atomic!` declarative macro stays `#[macro_export]`-mounted at
the crate root, co-located with the `atomic` fn for readability.

mod.rs: 4139 → 3852 LOC (-287). Both new files are sibling leaves
with zero cross-module dependencies beyond `super::{Pool, ExecError,
Dialect}` + the public `PoolTx` enum.

1482 lib tests pass on the sqlite,tenancy litmus.
